+++ src/main.rs
@@ -15,13 +15,13 @@ fn main() -> io::Result<()> {
     let train_images = train_data
         .images
         .iter()
-        .map(|vec| vec.iter().map(|val| *val as f64).collect())
+        .map(|vec| vec.iter().map(|val| *val as f32).collect())
         .collect();
     let train_labels = train_data
         .labels
         .iter()
         .map(|val| {
-            let mut tmp = vec![0f64; 10];
+            let mut tmp = vec![0f32; 10];
             tmp[*val as usize] = 1.0;
             tmp
         })
@@ -36,16 +36,16 @@ fn main() -> io::Result<()> {
         "mnist/t10k-labels.idx1-ubyte",
     )?;
 
-    let test_images: Vec<Vec<f64>> = test_data
+    let test_images: Vec<Vec<f32>> = test_data
         .images
         .iter()
-        .map(|vec| vec.iter().map(|val| *val as f64).collect())
+        .map(|vec| vec.iter().map(|val| *val as f32).collect())
         .collect();
-    let test_labels: Vec<Vec<f64>> = test_data
+    let test_labels: Vec<Vec<f32>> = test_data
         .labels
         .iter()
         .map(|val| {
-            let mut tmp = vec![0f64; 10];
+            let mut tmp = vec![0f32; 10];
             tmp[*val as usize] = 1.0;
             tmp
         })
